Added the version of the dependencies on the main project. Fixed unused dependencies, and integrate them all on the parent project when was possible. Replaced the common-lang3 with org.springframework for inheritance field retrieval. Updated the version to the latest mule-agent-api.
Downgrade the commons-lang3 to commons-lang which is the one used on Mule.

Fixed dependencies that aren't provided.
Fixed code line length.
Fixed a bug on template making.

Removed the assembly generation.

Fixed the pom dependencies, and the final assembly name.

Added support for the Custom Business Events.

Better handling of the annotations, and custom business events.

Removed last blank space.

Fixed the formatted date output.

Added javax validation.

Refactored the serialization of the different value types.

Refactoring of the DB Internal Handler in order to allow the persistence of the related entities.
Added the UUID time based generator for better performance by reducing the indexing fragmentation.

Added business events.

Added the commons logging dependency, since OneOffSwitch uses it.

Changed the UUID data type to CHAR which is more simple for different RDBMS to store.
Added the executeBatch() for the statements.
Removed the event name from the business, since it will be on the Map collection.

Fixed the UUUID generator dependency.

Added the Custom Business Events to the test.

Added the Custom Business Events and Annotations table documentation.

Fixed the documentation regarding the pattern usage.

Fixed formatting styles.